Remove unused imports and refactor to separated files
