+++ src/main/java/com/mercadolibre/bootcampmelifrescos/dtos/SectionDTO.java
@@ -1,0 +1,11 @@
+package com.mercadolibre.bootcampmelifrescos.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+class SectionDTO {
+    private Long sectionCode;
+    private Long warehouseCode;
+}
