+++ src/main/java/com/mercadolibre/bootcampmelifrescos/dtos/request/InboundOrderRequest.java
@@ -1,0 +1,11 @@
+package com.mercadolibre.bootcampmelifrescos.dtos.request;
+
+import com.mercadolibre.bootcampmelifrescos.dtos.InboundOrderDTO;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+public class InboundOrderRequest {
+    private InboundOrderDTO inboundOrder;
+}
